+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Initiated Development
 
+- Added Utils for handling FTP file ingestion - `SCT-004`
+
 - Added template for Python Data Structures data ingestion - `SCT-003`
 
 - Added template for RestAPI data ingestion - `SCT-001`